implemented suspend/resume
Attempted to implement the suspend and resume methods. Project still currently fails 10/16 of the grading tests. It is passing all six of the basic region tests. Also added findSmallest and findLargest helper methods to help find the largest or smallest coordinate of changed pixels in order to try to make the union of all the changed pixels.